requirements.txt
requests
boto3
python-dotenv
urllib3
tqdm 